requirements.txt
crispy-bootstrap5==0.6
Django==4.1.1
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
whitenoise==6.2.0
psycopg2-binary==2.9.3

pip~=22.3.1
attrs~=22.1.0
wheel~=0.38.4
cryptography~=39.0.0
Jinja2~=3.1.2
setuptools~=65.5.1
future~=0.18.2
six~=1.16.0